Created a dialog to prevent close without saving
git-svn-id: https://svn.forge.morfeo-project.org/fast-fp7project@661 d8b0c59b-6544-0410-94d9-83a5658a058e

+++ trunk/GVS/media/js/gvs/documents/ScreenDocument.js
@@ -253,28 +253,30 @@ var ScreenDocument = Class.create(PaletteDocument,
         };
     },
 
-
     /**
      * @private
      * @override
      */
-    _closeDocument: function($super) {
-        $super();
-        this._description.getCanvasInstances().each(function(instance) {
-            instance.destroyWires();
-        }.bind(this));
+    _effectiveCloseDocument: function($super, /** String */ status) {
+        $super(status);
 
-        this._description.getConditionInstances().each(function(instance) {
-            instance.destroyWires();
-        }.bind(this));
+        if (!status || status == ConfirmDialog.DISCARD) {
+            this._description.getCanvasInstances().each(function(instance) {
+                instance.destroyWires();
+            }.bind(this));
 
-        this._dojoConnections.each(function(connection){
-            dojo.disconnect(connection);
-        }.bind(this));
+            this._description.getConditionInstances().each(function(instance) {
+                instance.destroyWires();
+            }.bind(this));
 
-        // Remove circular dependencies
-        this._dojoConnections = null;
-        this._repaint = null;
+            this._dojoConnections.each(function(connection){
+                dojo.disconnect(connection);
+            }.bind(this));
+
+            // Remove circular dependencies
+            this._dojoConnections = null;
+            this._repaint = null;
+        }      
     },
